Update JobSubmitter interface to take in clusterName and return submission status (#697)
## Summary
This is needed for agent to be able to track status of submitted jobs
and report them back to the orchestration service

## Checklist
- [ ] Added Unit Tests
- [x] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update

+++ api/thrift/agent.thrift
@@ -5,8 +5,8 @@ include "common.thrift"
 struct YarnAutoScalingSpec {
     1: optional i32 minInstances
     2: optional i32 maxInstances
-    3: optional i32 scaleUpFactor // 1.5x, 2x etc
-    4: optional i32 scaleDownFactor
+    3: optional double scaleUpFactor // 1.5x, 2x etc
+    4: optional double scaleDownFactor
     5: optional string cooldownPeriod
 }
